renamed entityObject to ObjectEntity
